Mudando a organização das pastas e o nome de alguns arquivos.
