+++ API/OtherSolutions/CCEARc/Integrations/Events/GroupCcearcContractsReajustedIntegrationEvent.cs
@@ -1,0 +1,12 @@
+﻿namespace API.OtherSolutions.CCEARc.Integrations.Events
+{
+    public class GroupCcearcContractsReajustedIntegrationEvent
+    {
+        public GroupCcearcContractsReajustedIntegrationEvent(List<string> contracts)
+        {
+            Contracts = contracts;
+        }
+
+        public List<string> Contracts { get; private set; }
+    }
+}
