--- API/OtherSolutions/CCEARc/IntegrationsEvents/CcearcContractReajustedIntegrationEvent.cs
@@ -1,12 +1,0 @@
-﻿namespace API.OtherSolutions.CCEARc.IntegrationsEvents
-{
-    public class CcearcContractReajustedIntegrationEvent
-    {
-        public CcearcContractReajustedIntegrationEvent(string contract)
-        {
-            Contract = contract;
-        }
-
-        public string Contract { get; private set; }
-    }
-}
