--- API/OtherSolutions/CCEARc/IntegrationsEvents/GroupCcearcContractsReajustedIntegrationEvent.cs
@@ -1,12 +1,0 @@
-﻿namespace API.OtherSolutions.CCEARc.IntegrationsEvents
-{
-    public class GroupCcearcContractsReajustedIntegrationEvent
-    {
-        public GroupCcearcContractsReajustedIntegrationEvent(List<string> contracts)
-        {
-            Contracts = contracts;
-        }
-
-        public List<string> Contracts { get; private set; }
-    }
-}
