+++ API/OtherSolutions/CCEARc/Internal/Commands/GroupCcearcContractsInternalCommand.cs
@@ -1,0 +1,12 @@
+﻿namespace API.OtherSolutions.CCEARc.Internal.Commands
+{
+    public class GroupCcearcContractsInternalCommand
+    {
+        public GroupCcearcContractsInternalCommand(List<string> contracts)
+        {
+            Contracts = contracts;
+        }
+
+        public List<string> Contracts { get; private set; }
+    }
+}
