+++ API/OtherSolutions/CCEARc/Internal/Commands/StartSagaForAjusteContractPriceForUpdateFinanceIndexInternalCommand.cs
@@ -1,0 +1,12 @@
+﻿namespace API.OtherSolutions.CCEARc.Internal.Commands
+{
+    public class StartSagaForAjusteContractPriceForUpdateFinanceIndexInternalCommand
+    {
+        public string SagaId { get; set; }
+
+        public StartSagaForAjusteContractPriceForUpdateFinanceIndexInternalCommand()
+        {
+            SagaId = Guid.NewGuid().ToString();
+        }
+    }
+}
