--- API/OtherSolutions/CCEARc/InternalCommands/FinancialIndexValuesChangedInternalCommand.cs
@@ -1,12 +1,0 @@
-﻿namespace API.OtherSolutions.CCEARc.InternalCommands
-{
-    public class FinancialIndexValuesChangedInternalCommand
-    {
-        public string SagaId { get; set; }
-
-        public FinancialIndexValuesChangedInternalCommand()
-        {
-            SagaId = Guid.NewGuid().ToString();
-        }
-    }
-}
