+++ API/OtherSolutions/CCEARc/Saga/FinancialIndexUpdatedSaga.cs
@@ -1,18 +1,17 @@
-﻿using API.OtherSolutions.CCEARc.IntegrationsCommands;
-using API.OtherSolutions.CCEARc.InternalCommands;
-using API.OtherSolutions.CCEARc.InternalEvents;
-using API.OtherSolutions.EnergyBalance.IntegrationsCommands;
+﻿using API.OtherSolutions.CCEARc.Integrations.Commands;
+using API.OtherSolutions.CCEARc.Internal.Commands;
+using API.OtherSolutions.CCEARc.Internal.Events;
+using API.OtherSolutions.EnergyBalance.Integrations.Events;
 using Rebus.Bus;
 using Rebus.Handlers;
-using Rebus.Messages;
 using Rebus.Sagas;
 
 namespace API.OtherSolutions.CCEARc.Saga
 {
     public class FinancialIndexUpdatedSaga :
             Saga<FinancialIndexUpdateSagaData>,
-            IAmInitiatedBy<FinancialIndexValuesChangedInternalCommand>,
-            IHandleMessages<CcearcContractForReajustmentCommandIntegration>,
+            IAmInitiatedBy<StartSagaForAjusteContractPriceForUpdateFinanceIndexInternalCommand>,
+            IHandleMessages<GetedRelationTheContractsForAjusteThePriceForUpdateFinanceIndexIntegrationEvent>,
             IHandleMessages<CcearcContractReajustedForIndexInternalEvent>        
     {
         private readonly IBus _bus;
@@ -22,13 +21,13 @@ namespace API.OtherSolutions.CCEARc.Saga
             _bus = bus;
         }
 
-        public Task Handle(FinancialIndexValuesChangedInternalCommand message)
+        public Task Handle(StartSagaForAjusteContractPriceForUpdateFinanceIndexInternalCommand message)
         {
-            _bus.Publish(new CalculateCcearcContractsForIndexReadjustmentCommandIntegration(message.SagaId));
+            _bus.Publish(new CalculateTheCcearcContractsAjustedPriceForUpdateFinanceIndexCommandIntegration(message.SagaId));
             return Task.CompletedTask;
         }
 
-        public Task Handle(CcearcContractForReajustmentCommandIntegration message)
+        public Task Handle(GetedRelationTheContractsForAjusteThePriceForUpdateFinanceIndexIntegrationEvent message)
         {
             Data.ContratcToReajust = message.Contracts.ToList();
             return Task.CompletedTask;
@@ -47,8 +46,8 @@ namespace API.OtherSolutions.CCEARc.Saga
 
         protected override void CorrelateMessages(ICorrelationConfig<FinancialIndexUpdateSagaData> config)
         {
-            config.Correlate<FinancialIndexValuesChangedInternalCommand>(message => message.SagaId, d => d.CorrelationId);
-            config.Correlate<CcearcContractForReajustmentCommandIntegration>(message => message.SagaId, d => d.CorrelationId);
+            config.Correlate<StartSagaForAjusteContractPriceForUpdateFinanceIndexInternalCommand>(message => message.SagaId, d => d.CorrelationId);
+            config.Correlate<GetedRelationTheContractsForAjusteThePriceForUpdateFinanceIndexIntegrationEvent>(message => message.SagaId, d => d.CorrelationId);
             config.Correlate<CcearcContractReajustedForIndexInternalEvent>(message => message.SagaId, d => d.CorrelationId);
         }
     }
