--- API/OtherSolutions/EnergyBalance/IntegrationsCommands/CcearcContractForReajustmentCommandIntegration.cs
@@ -1,15 +1,0 @@
-﻿namespace API.OtherSolutions.EnergyBalance.IntegrationsCommands
-{
-    public class CcearcContractForReajustmentCommandIntegration
-    {
-        public CcearcContractForReajustmentCommandIntegration(string sagaId, IEnumerable<string> contracts)
-        {
-            SagaId = sagaId;
-            Contracts = contracts;
-        }
-
-        public string SagaId { get; set; }
-        public IEnumerable<string> Contracts { get; private set; }
-        
-    }
-}
