+++ API/OtherSolutions/EnergyBalance/Internal/Commands/CalculatePriceOperationsCcearcInternalCommand.cs
@@ -1,0 +1,14 @@
+﻿namespace API.OtherSolutions.EnergyBalance.Internal.Commands
+{
+    public class CalculatePriceOperationsCcearcInternalCommand
+    {
+        public CalculatePriceOperationsCcearcInternalCommand(string sagaId, string contractId)
+        {
+            SagaId = sagaId;
+            ContractId = contractId;
+        }
+
+        public string SagaId { get; private set; }
+        public string ContractId { get; private set; }
+    }
+}
