+++ API/OtherSolutions/Finance/Integrations/Events/FinanceIndexUpdatedIntegrationEvent.cs
@@ -1,0 +1,4 @@
+﻿namespace API.OtherSolutions.Finance.Integrations.Events
+{
+    public class FinanceIndexUpdatedIntegrationEvent { }
+}
