--- API/OtherSolutions/Finance/IntegrationsEvents/FinancialIndexValuesChangedIntegrationEvent.cs
@@ -1,6 +1,0 @@
-﻿namespace API.OtherSolutions.Finance.IntegrationsEvents
-{
-    public class FinancialIndexValuesChangedIntegrationEvent
-    {
-    }
-}
