+++ API/Program.cs
@@ -1,14 +1,13 @@
 
 using API;
-using API.OtherSolutions.CCEARc.IntegrationsCommands;
-using API.OtherSolutions.CCEARc.IntegrationsEvents;
-using API.OtherSolutions.CCEARc.IntegrationsHandlers;
-using API.OtherSolutions.CCEARc.InternalCommands;
-using API.OtherSolutions.CCEARc.InternalEvents;
-using API.OtherSolutions.EnergyBalance.IntegrationsCommands;
-using API.OtherSolutions.EnergyBalance.IntegrationsHandlers;
-using API.OtherSolutions.EnergyBalance.InternalCommand;
-using API.OtherSolutions.Finance.IntegrationsEvents;
+using API.OtherSolutions.CCEARc.Integrations.Commands;
+using API.OtherSolutions.CCEARc.Integrations.Events;
+using API.OtherSolutions.CCEARc.Internal.Commands;
+using API.OtherSolutions.CCEARc.Internal.Events;
+using API.OtherSolutions.EnergyBalance.Integrations.Commands;
+using API.OtherSolutions.EnergyBalance.Integrations.Events;
+using API.OtherSolutions.EnergyBalance.Internal.Commands;
+using API.OtherSolutions.Finance.Integrations.Events;
 using Rebus.Persistence.InMem;
 using Rebus.Routing.TypeBased;
 using Rebus.ServiceProvider;
@@ -52,10 +51,25 @@ if (app.Environment.IsDevelopment())
 
 app.UseRebus(c =>
 {
-    c.Subscribe<FinancialIndexValuesChangedIntegrationEvent>().Wait();//Recebendo indice do financeiro - ajuste
-    c.Subscribe<FinancialIndexValuesChangedInternalCommand>().Wait();//Realizando checagem ccearc e iniciando a saga
-    c.Subscribe<CalculateCcearcContractsForIndexReadjustmentCommandIntegration>().Wait();//Disparando evento para o Energy Balance
-    c.Subscribe<CcearcContractForReajustmentCommandIntegration>().Wait();//Recebendo lista com os ids dos contratos que serão recalculados no EnergyBalance
+    c.Subscribe<FinanceIndexUpdatedIntegrationEvent>().Wait();
+    //Endpoint do financeiro para atualização de indice é acionado.
+    //Financeiro dispara o evento informando que o indice foi atualizado.
+    //Ccearc recebe o evento para inicio do processo de atualização dos preços dos contratos.
+    
+    c.Subscribe<StartSagaForAjusteContractPriceForUpdateFinanceIndexInternalCommand>().Wait();
+    //Ccearc faz a checagem se tem contratos que tem que ter o preço atualizado.
+    //Se tiver, dispara o command para inicio da Saga.
+
+
+    c.Subscribe<CalculateTheCcearcContractsAjustedPriceForUpdateFinanceIndexCommandIntegration>().Wait();
+    //Ccearc dispara o comando para calcular o preço ajustado dos contratos pela atualização de index financeiro.
+    //EnergyBalance recebe o comando.
+
+
+    c.Subscribe<GetedRelationTheContractsForAjusteThePriceForUpdateFinanceIndexIntegrationEvent>().Wait();
+    //EnergyBalance obtem os contratos que terão os preços ajustados e dispara o evento com o id deles.
+    //Ccearc recebe o evento e atualiza na Saga a lista dos contratos a serem ajustados.
+
     c.Subscribe<CalculatePriceOperationsCcearcInternalCommand>().Wait();//Disparando o comando interno no EnergyBalance para calular o preco dos contratos.
     c.Subscribe<PriceCcearcReajustedIntegrationCommand>().Wait();//Disparando do Energy Balance para o Ccearc o evento com o contrato com o preço ajustado.
     c.Subscribe<CcearcContractReajustedForIndexInternalEvent>().Wait();//Recebendo o contrato que foi ajustado internamente e enviado para o BI
